pyproject.toml
[project]
name = "fastapi-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "httptools>=0.6.4",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
]
